Split register page and login page
