--- back/data/index.js
@@ -1,112 +1,0 @@
-import mongoose from "mongoose";
-
-const userIds = [
-  new mongoose.Types.ObjectId(),
-  new mongoose.Types.ObjectId(),
-  new mongoose.Types.ObjectId(),
-  new mongoose.Types.ObjectId(),
-  new mongoose.Types.ObjectId(),
-  new mongoose.Types.ObjectId(),
-  new mongoose.Types.ObjectId(),
-  new mongoose.Types.ObjectId(),
-];
-
-export const users = [
-  {
-    _id: userIds[0],
-    firstName: "test",
-    lastName: "me",
-    email: "aaaaaaa@gmail.com",
-    password: "$2b$10$dsasdgsagasda//G9JxQ4bQ8KXf4OAIe/X/AK9skyWUy",
-    picturePath: "p11.jpeg",
-    occupation: "Software Engineer",
-    createdAt: 1115211422,
-    updatedAt: 1115211422,
-    __v: 0,
-  },
-  {
-    _id: userIds[1],
-    firstName: "Steve",
-    lastName: "Ralph",
-    email: "thataaa@gmail.com",
-    password: "$!FEAS@!O)_IDJda//G9JxQ4bQ8KXf4OAIe/X/AK9skyWUy",
-    picturePath: "p3.jpeg",
-    occupation: "Degenerate",
-    createdAt: 1595589072,
-    updatedAt: 1595589072,
-    __v: 0,
-  },
-  {
-    _id: userIds[2],
-    firstName: "Some",
-    lastName: "Guy",
-    email: "someguy@gmail.com",
-    password: "da39a3ee5e6b4b0d3255bfef95601890afd80709",
-    picturePath: "p4.jpeg",
-    occupation: "Data Scientist Hacker",
-    createdAt: 1288090662,
-    updatedAt: 1288090662,
-    __v: 0,
-  },
-  {
-    _id: userIds[3],
-    firstName: "Whatcha",
-    lastName: "Doing",
-    email: "whatchadoing@gmail.com",
-    password: "$2b$10$dsasdgsagasda//G9JxQ4bQ8KXf4OAIe/X/AK9skyWUy",
-    picturePath: "p6.jpeg",
-    occupation: "Educator",
-    createdAt: 1219214568,
-    updatedAt: 1219214568,
-    __v: 0,
-  },
-  {
-    _id: userIds[4],
-    firstName: "Jane",
-    lastName: "Doe",
-    email: "janedoe@gmail.com",
-    password: "$2b$10$dsasdgsagasda//G9JxQ4bQ8KXf4OAIe/X/AK9skyWUy",
-    picturePath: "p5.jpeg",
-    occupation: "Hacker",
-    createdAt: 1493463661,
-    updatedAt: 1493463661,
-    __v: 0,
-  },
-  {
-    _id: userIds[5],
-    firstName: "Harvey",
-    lastName: "Dunn",
-    email: "harveydunn@gmail.com",
-    password: "$2b$10$dsasdgsagasda//G9JxQ4bQ8KXf4OAIe/X/AK9skyWUy",
-    picturePath: "p7.jpeg",
-    occupation: "Journalist",
-    createdAt: 1381326073,
-    updatedAt: 1381326073,
-    __v: 0,
-  },
-  {
-    _id: userIds[6],
-    firstName: "Carly",
-    lastName: "Vowel",
-    email: "carlyvowel@gmail.com",
-    password: "$2b$10$dsasdgsagasda//G9JxQ4bQ8KXf4OAIe/X/AK9skyWUy",
-    picturePath: "p8.jpeg",
-    occupation: "Nurse",
-    createdAt: 1714704324,
-    updatedAt: 1642716557,
-    __v: 0,
-  },
-  {
-    _id: userIds[7],
-    firstName: "Jessica",
-    lastName: "Dunn",
-    email: "jessicadunn@gmail.com",
-    password: "$2b$10$dsasdgsagasda//G9JxQ4bQ8KXf4OAIe/X/AK9skyWUy",
-    picturePath: "p9.jpeg",
-    friends: [],
-    occupation: "A Student",
-    createdAt: 1369908044,
-    updatedAt: 1359322268,
-    __v: 0,
-  },
-];
