+++ front/src/App.js
@@ -13,6 +13,7 @@ import { useSelector } from "react-redux";
 import { CssBaseline, ThemeProvider } from "@mui/material";
 import { createTheme } from "@mui/material/styles";
 import { themeSettings } from "./theme";
+import RegisterPage from 'scenes/registerPage';
 
 function App() {
   const mode = useSelector((state) => state.mode);
@@ -26,6 +27,10 @@ function App() {
           <CssBaseline />
           <Routes>
             <Route path="/" element={<LoginPage />} />
+            <Route
+              path="/register"
+              element={isAuth ? <RegisterPage /> : <Navigate to="/" />}
+            />
             <Route
               path="/home"
               element={isAuth ? <HomePage /> : <Navigate to="/" />}
